rename approval to proposal
