+++ CLIENT/CLIENT.ProcureAccess/src/app/features/proposal/list/proposal-list.html
