+++ CLIENT/CLIENT.ProcureAccess/src/app/features/proposal/models/approve-proposal-request.model.ts
