+++ CLIENT/CLIENT.ProcureAccess/src/app/features/proposal/models/proposal.dto.ts
