+++ CLIENT/CLIENT.ProcureAccess/src/app/features/proposal/state/with-proposal.ts
